+++ python/valis/db/db.py
@@ -2,10 +2,11 @@
 # -*- coding: utf-8 -*-
 #
 
+import asyncio
 from contextvars import ContextVar
 
 import peewee
-from fastapi import HTTPException
+from fastapi import HTTPException, Depends
 from sdssdb.peewee.sdss5db import database as pdb
 from sdssdb.sqlalchemy.sdss5db import database as sdb
 
@@ -18,10 +19,14 @@ db_state_default = {"closed": None, "conn": None, "ctx": None, "transactions": N
 db_state = ContextVar("db_state", default=db_state_default.copy())
 
 
-def reset_db_state():
-    """ Sub-depdency for get_db that resets the context connection state """
-    pdb._state._state.set(db_state_default.copy())
-    pdb._state.reset()
+async def reset_db_state():
+    """ Sub-dependency for get_db that resets the context connection state """
+
+    from valis.main import settings
+
+    if settings.db_reset:
+        pdb._state._state.set(db_state_default.copy())
+        pdb._state.reset()
 
 
 class PeeweeConnectionState(peewee._ConnectionState):
@@ -64,16 +69,18 @@ def connect_db(db, orm: str = 'peewee'):
 
     return db
 
-def get_pw_db():
+async def get_pw_db(db_state=Depends(reset_db_state)):
     """ Dependency to connect a database with peewee """
 
     from valis.main import settings
 
-    if settings.db_reset:
-        reset_db_state()
-
     # connect to the db, yield None since we don't need the db in peewee
-    db = connect_db(pdb, orm='peewee')
+    if settings.db_reset:
+        db = connect_db(pdb, orm='peewee')
+    else:
+        async with asyncio.Lock():
+            db = connect_db(pdb, orm='peewee')
+
     try:
         yield db
     finally: