+++ python/valis/wsgi_conf.py
@@ -30,3 +30,4 @@ daemon = False
 errorlog = os.path.join(os.getenv("VALIS_LOGS_DIR", '/tmp/valis'), 'valis_app_error.log')
 accesslog = os.path.join(os.getenv("VALIS_LOGS_DIR", '/tmp/valis'), 'valis_app_access.log')
 root_path = '/valis'
+timeout = 600